Better diagnostic message when target is not a Moose class
If the target of the composed behavior is not a Moose class, the call
to 'apply_all_roles' will fail with a bizarre error message.  This
cann occur if, for example, your "use Moose" declaration follows the
"use MooseX::ComposedBehavior" declaration in the target file.

Moose itself should diagnose this more clearly, but in the meantime we
add a diagnosis here.